% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{app_dictionary}
\alias{app_dictionary}
\title{dataset: App Dictionary}
\format{
A list of named dataframes
}
\usage{
app_dictionary
}
\value{
A list of named dataframes
}
\description{
An list of tables associated with application mapping for pretty GUI and selection
}
\keyword{datasets}
