% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_index.R
\name{build_index}
\alias{build_index}
\title{Build_Index}
\usage{
build_index(sub = "", file_path, api_key)
}
\arguments{
\item{sub}{a string containing the target subdirectory beneath https://api.eia.gov/v2/ (i.e. "petroleum/sum")}

\item{file_path}{a string with a .rds extension containing the destination path for the completed index, cached locally}

\item{api_key}{a string containing your eia api key}
}
\description{
Intended for use after a metadata tree has been created using route_tree(). Not currently implemented
}
