% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_data_types.R
\name{by_data_types}
\alias{by_data_types}
\title{by_data_types}
\usage{
by_data_types(dindex_in, selections)
}
\arguments{
\item{dindex_in}{A data_index subset dataframe}

\item{selections}{A list/vector of data types (strings)}
}
\value{
A data_index dataframe, with narrowed data selections
}
\description{
EIA API calls will not return endpoint rows unless all specified/queried data types are represented in the returned rows. As such, sometimes removing/narrowing the data request is required in order to ensure a valid response. This function simplifies the process of narrowing your request. Use in tandem with show_data_types()
}
