% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_index.R
\name{collapse_index}
\alias{collapse_index}
\title{Collapse Index
A function for converting lists of dataframe to a row stacked dataframe}
\usage{
collapse_index(list_in)
}
\arguments{
\item{list_in}{A list of dataframes.}
}
\value{
A dataframe
}
\description{
Collapse Index
A function for converting lists of dataframe to a row stacked dataframe
}
