% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_index}
\alias{data_index}
\title{dataset: Data Index}
\format{
A list of named dataframes
}
\usage{
data_index
}
\value{
A list of named dataframes
}
\description{
An index of eia data series, isolated by route_1_id to improve maintainability. Used in tandem with dindex_get_data().
}
\keyword{datasets}
