% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_routes.R
\name{detect_routes}
\alias{detect_routes}
\title{Detect routes}
\usage{
detect_routes(m_data)
}
\arguments{
\item{m_data}{A list containing eia metadata}
}
\value{
Logical
}
\description{
Determine if the routes slot is present in metadata
}
