% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_call.R
\name{eia_call}
\alias{eia_call}
\title{eia_call}
\usage{
eia_call(endpoint)
}
\arguments{
\item{endpoint}{a url endpoint, with headers attached}
}
\value{
A response in json list format
}
\description{
eia_call
}
