% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_data.R
\name{eia_data}
\alias{eia_data}
\title{eia_data_call}
\usage{
eia_data(
  api_endpoint,
  freq,
  facets = NULL,
  start = NA,
  end = NA,
  sort_df = data.frame(sortby = c("period"), direction = c("desc")),
  data_types = NULL,
  offset = NA,
  api_key
)
}
\arguments{
\item{api_endpoint}{an eia api endpoint subsirectory, as a string}

\item{freq}{a valid eia frequency for the endpoint, a string}

\item{facets}{a dataframe with cols facet_name, facet_id}

\item{start}{A string in "YYYY-MM-DD" format}

\item{end}{A string in "YYYY-MM-DD" format}

\item{sort_df}{A dataframe with sortby = column_name and direction = c(asc,desc)columns. First rows take sort priority}

\item{data_types}{A list of valid datatypes to pull}

\item{offset}{A numeric, for pagination of results}

\item{api_key}{Your eia api key}
}
\value{
A dataframe
}
\description{
eia_data_call
}
