% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_map.R
\name{eia_map}
\alias{eia_map}
\title{EIA Mapping Request}
\usage{
eia_map(sub, offset, freq, api_key)
}
\arguments{
\item{sub}{String containing subdirectory for EIA endpoint}

\item{offset}{Used to offset pagination in sets of 5000}

\item{freq}{String with an EIA frequency code}

\item{api_key}{String with your EIA API key}
}
\value{
A DataFrame containing the unique identifiers for updated/recent data for products
}
\description{
EIA Mapping Request
}
