% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_meta.R
\name{eia_meta}
\alias{eia_meta}
\title{EIA Metadata Request}
\usage{
eia_meta(sub, api_key)
}
\arguments{
\item{sub}{A string containing subdirectory, or tree of subdir filepaths for the eia API endpoint}

\item{api_key}{String with your eia api key}
}
\value{
A list containing the json data
}
\description{
EIA Metadata Request
}
