% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_reports.R
\name{eia_reports}
\alias{eia_reports}
\title{EIA Generate All Weekly Status Reports}
\usage{
eia_reports(category)
}
\arguments{
\item{category}{String listing report name, options('petroleum')}
}
\value{
DataFrame
}
\description{
EIA Generate All Weekly Status Reports
}
