% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_today_in_energy.R
\name{eia_today_in_energy}
\alias{eia_today_in_energy}
\title{EIA Today In Energy Reports List}
\usage{
eia_today_in_energy()
}
\value{
A list of strings containing titles, url links, and dates of recent publications.
}
\description{
EIA Today In Energy Reports List
}
