% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_version.R
\name{eia_version}
\alias{eia_version}
\title{EIA Version Request}
\usage{
eia_version(api_key)
}
\arguments{
\item{api_key}{A string containing your EIA API Key}
}
\value{
A list of strings containing API and Excel Versions for EIA's web service
}
\description{
EIA Version Request
}
