% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiatools-package.R
\docType{package}
\name{eiatools-package}
\alias{eiatools}
\alias{eiatools-package}
\title{eiatools: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Justin Powley \email{jpowley@ualberta.ca}

Authors:
\itemize{
  \item Brayden Boyko \email{braydenboyko@boykowealth.com}
}

}
\keyword{internal}
