% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_freq.R
\name{get_all_freq}
\alias{get_all_freq}
\title{get_all_freq}
\usage{
get_all_freq(m_data)
}
\arguments{
\item{m_data}{A list with eia metadata}
}
\value{
A vector of frequency codes applicable to the eia api.
}
\description{
get_all_freq
}
