% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_types.R
\name{get_data_types}
\alias{get_data_types}
\title{get_data_types}
\usage{
get_data_types(m_data)
}
\arguments{
\item{m_data}{A list with eia metadata}
}
\value{
A vector of data codes applicable to the eia api.
}
\description{
get_data_types
}
