% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facet_data.R
\name{get_facet_data}
\alias{get_facet_data}
\title{get_facet_data}
\usage{
get_facet_data(sub, facet_id, api_key)
}
\arguments{
\item{sub}{a string}

\item{facet_id}{a string}

\item{api_key}{your eia api key}
}
\value{
A dataframe
}
\description{
get_facet_data
}
