% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facet_types.R
\name{get_facet_types}
\alias{get_facet_types}
\title{get_facet_types}
\usage{
get_facet_types(m_data)
}
\arguments{
\item{m_data}{A list with eia metadata}
}
\value{
A vector of facet codes applicable to the eia api.
}
\description{
get_facet_types
}
