% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reports.R
\name{get_reports}
\alias{get_reports}
\title{EIA Get Weekly Status Report Request}
\usage{
get_reports(url)
}
\arguments{
\item{url}{String with the URL to given xlsx file}
}
\value{
DataFrame
}
\description{
EIA Get Weekly Status Report Request
}
