% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_routes.R
\name{get_routes}
\alias{get_routes}
\title{get routes
gets the routes dataframe from eia metadata}
\usage{
get_routes(m_data)
}
\arguments{
\item{m_data}{list with eia metadata}
}
\value{
A dataframe
}
\description{
get routes
gets the routes dataframe from eia metadata
}
