% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_reader.R
\name{html_reader}
\alias{html_reader}
\title{eia News Report Reader}
\usage{
html_reader(url)
}
\arguments{
\item{url}{The URL of report wanting to read.}
}
\value{
HTML code for specified report, can be rendered in ShinyApp
}
\description{
eia News Report Reader
}
