% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_headers.R
\name{map_headers}
\alias{map_headers}
\title{Remap Headers}
\usage{
map_headers(url, api_key, headers = NULL)
}
\arguments{
\item{url}{A string with the url endpoint}

\item{api_key}{A string with your EIA API key}

\item{headers}{A named list(sort = "", direction = "", data = "", freq = ""...) with header names and arguments}
}
\value{
A string url with headers mapped
}
\description{
Remap Headers
}
