% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_data}
\alias{my_data}
\title{dataset: my_data}
\format{
A string
}
\usage{
my_data
}
\value{
A string
}
\description{
A test data point with a cheeky comment about our packages current needs.
}
\keyword{datasets}
