% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route_index}
\alias{route_index}
\title{dataset: Route Index}
\format{
A list of named dataframes
}
\usage{
route_index
}
\value{
A list of named dataframes
}
\description{
An index of eia routes, isolated by route_1_id to improve maintainability.
}
\keyword{datasets}
