% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_tree.R
\name{route_tree}
\alias{route_tree}
\title{EIA Meta Data Tree}
\usage{
route_tree(sub = "", api_key, iter = 1, iter_offset = 1)
}
\arguments{
\item{sub}{A subdirectory on EIA's api url, for example, "petroleum"}

\item{api_key}{Your eia api key}

\item{iter}{A counter used in recursion, internal to the function}

\item{iter_offset}{A reindexing value used in recursion, internal to the function}
}
\value{
A metadata dataframe
}
\description{
EIA Meta Data Tree
}
