% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route_tree_sample}
\alias{route_tree_sample}
\title{dataset: route_tree_sample}
\format{
A dataframe
}
\usage{
route_tree_sample
}
\value{
A dataframe
}
\description{
Example of output from route_tree function targeting "petroleum/sum". This function will later be for internal use within the package, defining endpoints for future functions.
}
\keyword{datasets}
