% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data_types.R
\name{show_data_types}
\alias{show_data_types}
\title{show_data_types}
\usage{
show_data_types(dindex_in)
}
\arguments{
\item{dindex_in}{A data_index dataframe}
}
\value{
A vector of data types
}
\description{
Shows the various datatypes available within a data_index dataframe
}
