% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dindex_get_data.R
\name{dindex_get_data}
\alias{dindex_get_data}
\title{d(ata)index_get_data}
\usage{
dindex_get_data(dindex_table, api_key, start = NA, end = NA)
}
\arguments{
\item{dindex_table}{a subset of the data_index table}

\item{api_key}{your eia api key}

\item{start}{"YYYY-MM-DD"}

\item{end}{"YYYY-MM-DD"}
}
\value{
A dataframe
}
\description{
A function accepting rows from the data_index object and then fetches from EIA
}
