% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_call.R
\name{eia_call}
\alias{eia_call}
\title{eia_call}
\usage{
eia_call(endpoint, sleep = 5)
}
\arguments{
\item{endpoint}{a url endpoint, with headers attached}

\item{sleep}{integer of number of seconds to wait before continuing}
}
\value{
A response in json list format
}
\description{
Function for handling api requests at an atomic level.
Also includes url cleaning components autaomatically applied by most web broswers
}
