% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_map.R
\name{eia_map}
\alias{eia_map}
\title{EIA Mapping Request}
\usage{
eia_map(sub, offset, freq, api_key)
}
\arguments{
\item{offset}{Used to offset pagination in sets of 5000}

\item{api_key}{String with your eia API key}

\item{type}{String with URL extension for product type}
}
\value{
A DataFrame containing the unique identifiers for all products
}
\description{
EIA Mapping Request
}
