% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiatools-package.R
\docType{package}
\name{eiatools-package}
\alias{eiatools}
\alias{eiatools-package}
\title{eiatools: What the Package Does (One Line, Title Case)}
\description{
EIATools is an endpoint archival tool, and api wrapper for the US Energy Information Administration. It's function is to provide a development toolkit for integrating both tools drawing directly from EIA data, and tools responsible for archiving and updating EIA's available datasets as they become available.
}
\author{
\strong{Maintainer}: Justin Powley \email{jpowley@ualberta.ca}

Authors:
\itemize{
  \item Brayden Boyko \email{braydenboyko@boykowealth.com}
}

}
\keyword{internal}
