% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_headers.R
\name{map_headers}
\alias{map_headers}
\title{Remap Headers}
\usage{
map_headers(url, api_key, headers = NULL)
}
\arguments{
\item{api_key}{A string}

\item{headers}{A named list with keys matching facet type.}
}
\value{
A string url with headers mapped
}
\description{
Remap Headers
}
